# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill, plus the sensitive
    config:edit_admin and chat:admin sub-scopes. The config scope also grants bot
    and provider access, but no parent scope grants a sensitive sub-scope.
...
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/stats:
    get:
      tags: [Stats]
      summary: Get runtime statistics
      operationId: getStats
      x-astrbot-scope: system
      parameters:
        - name: offset_sec
          in: query
          schema:
            type: integer
            default: 86400
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/stats/provider-tokens:
    get:
      tags: [Stats]
      summary: Get provider token usage statistics
      operationId: getProviderTokenStats
      x-astrbot-scope: system
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/stats/plugin-tokens:
    get:
      tags: [Stats]
      summary: Get reported token usage grouped by plugin
      operationId: getPluginTokenStats
      x-astrbot-scope: system
      parameters:
        - name: days
          in: query
          description: Supports 1, 3, or 7 days; other integers use 1 day.
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Plugin token usage in the selected time window
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: "#/components/schemas/PluginTokenStatsData"
...
  /api/v1/stats/first-notice:
    get:
      tags: [Stats]
      summary: Get first-run dashboard notice content
      operationId: getFirstNotice
      x-astrbot-scope: system
      parameters:
        - name: locale
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/stats/ghproxy/test:
    post:
      tags: [Stats]
      summary: Test a GitHub proxy endpoint
      operationId: testGhproxyConnection
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GhproxyTestRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/changelogs/{version}:
    get:
      tags: [Stats]
      summary: Get changelog content for a version
      operationId: getChangelog
      x-astrbot-scope: system
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/v1/stats/storage/cleanup:
    post:
      tags: [Stats]
      summary: Clean storage
      operationId: cleanupStorage
      x-astrbot-scope: system
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
    Error:
      description: Standard AstrBot error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    PluginTokenStatsData:
      type: object
      required: [days, items]
      properties:
        days:
          type: integer
          enum: [1, 3, 7]
        items:
          type: array
          items:
            type: object
            required: [plugin_id, display_name, token_input_other, token_input_cached, token_output, total_tokens]
            properties:
              plugin_id:
                type: string
              display_name:
                type: string
              token_input_other:
                type: integer
              token_input_cached:
                type: integer
              token_output:
                type: integer
              total_tokens:
                type: integer
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          const: error
        message:
          type: string
        data: {}
      additionalProperties: false
    DynamicConfig:
      type: object
      additionalProperties: true
...
    GhproxyTestRequest:
      type: object
      required: [proxy_url]
      properties:
        proxy_url:
          type: string
      additionalProperties: false
...